+++ components/delay/delay.go
@@ -7,6 +7,11 @@ import (
 	"github.com/ThreeDotsLabs/watermill/message"
 )
 
+// Delay represents a message's delay.
+// It can be either a delay until a specific time or a delay for a specific duration.
+// The zero value of Delay is a zero delay.
+//
+// IMPORTANT: Delay doesn't work with all Pub/Subs! Using it won't have any effect on Pub/Subs that don't support it.
 type Delay struct {
 	time     time.Time
 	duration time.Duration
@@ -16,6 +21,7 @@ func (d Delay) IsZero() bool {
 	return d.time.IsZero()
 }
 
+// Until returns a delay of the given time.
 func Until(delayedUntil time.Time) Delay {
 	return Delay{
 		time:     delayedUntil,
@@ -23,6 +29,7 @@ func Until(delayedUntil time.Time) Delay {
 	}
 }
 
+// For returns a delay of now plus the given duration.
 func For(delayedFor time.Duration) Delay {
 	return Delay{
 		time:     time.Now().UTC().Add(delayedFor),
@@ -33,62 +40,26 @@ func For(delayedFor time.Duration) Delay {
 type contextKey string
 
 var (
-	delayCtxKey = contextKey("delay")
+	delayContextKey = contextKey("delay")
 )
+
+// WithContext returns a new context with the given delay.
+// If used together with a publisher wrapped with NewPublisher, the delay will be applied to the message.
+//
+// IMPORTANT: Delay doesn't work with all Pub/Subs! Using it won't have any effect on Pub/Subs that don't support it.
+func WithContext(ctx context.Context, delay Delay) context.Context {
+	return context.WithValue(ctx, delayContextKey, delay)
+}
 
 const (
 	DelayedUntilKey = "delayed_until"
 	DelayedForKey   = "delayed_for"
 )
 
-func WithContext(ctx context.Context, delay Delay) context.Context {
-	return context.WithValue(ctx, delayCtxKey, delay)
-}
-
+// Message sets the delay metadata on the message.
+//
+// IMPORTANT: Delay doesn't work with all Pub/Subs! Using it won't have any effect on Pub/Subs that don't support it.
 func Message(msg *message.Message, delay Delay) {
 	msg.Metadata.Set(DelayedUntilKey, delay.time.Format(time.RFC3339))
 	msg.Metadata.Set(DelayedForKey, delay.duration.String())
-}
-
-type PublisherConfig struct {
-	DefaultDelay Delay
-}
-
-func NewPublisher(pub message.Publisher, config PublisherConfig) (message.Publisher, error) {
-	return &publisher{
-		pub:    pub,
-		config: config,
-	}, nil
-}
-
-type publisher struct {
-	pub    message.Publisher
-	config PublisherConfig
-}
-
-func (p *publisher) Publish(topic string, messages ...*message.Message) error {
-	for i := range messages {
-		p.applyDelay(messages[i])
-	}
-	return p.pub.Publish(topic, messages...)
-}
-
-func (p *publisher) Close() error {
-	return p.pub.Close()
-}
-
-func (p *publisher) applyDelay(msg *message.Message) {
-	if msg.Metadata.Get(DelayedForKey) != "" {
-		return
-	}
-
-	if msg.Context().Value(delayCtxKey) != nil {
-		delay := msg.Context().Value(delayCtxKey).(Delay)
-		Message(msg, delay)
-		return
-	}
-
-	if !p.config.DefaultDelay.IsZero() {
-		Message(msg, p.config.DefaultDelay)
-	}
 }
